### 2019-07-20 #1585 @whedon
[PRE REVIEW]: ORIS: An interactive software tool for prediction of replication origin in prokaryotic genomes
**Submitting author:** @urmi-21 (<a href="http://orcid.org/0000-0003-3703-0820">Urminder Singh</a>)
**Repository:** <a href="https://github.com/urmi-21/ORIS" target ="_blank">https://github.com/urmi-21/ORIS</a>
**Version:** v1.0
