### 2019-07-20 #1585 @danielskatz
👋 @urmi-21 - A fix for the paper compilation issues might be in https://github.com/urmi-21/ORIS/pull/2  Please merge this and see if it fixes the problem, by adding a comment here that says ```@whedon generate pdf``` - feel free to iterate
