### 2019-07-20 #1585 @urmi-21
@danielskatz Thanks for the suggesting the fix in paper.md. The compiled pdf looks good to me.
