### 2019-07-20 #1585 @danielskatz
While we are in the process of assigning an editor, you might suggestion some possible reviewers, either from [this list](https://bit.ly/joss-reviewers) (please start at the bottom of the list) or elsewhere, who might be able to review your submission. Also, feel free to suggest individuals who are not on this list by mentioning their GitHub handles here.
