### 2019-07-21 #1585 @csoneson
@holtgrewe, @jdeligt - would you be interested in reviewing this submission for JOSS?
