### 2019-07-21 #1585 @jdeligt
Count me in.

On Sun, 21 Jul 2019, 7:24 pm Charlotte Soneson, <notifications@github.com>
wrote:

> @holtgrewe <https://github.com/holtgrewe>, @jdeligt
> <https://github.com/jdeligt> - would you be interested in reviewing this
> submission for JOSS?
>
