+++ paper/paper.bib
@@ -2,7 +2,9 @@
   title={Essential cell biology},
   author={Alberts, Bruce and Bray, Dennis and Hopkin, Karen and Johnson, Alexander D and Lewis, Julian and Raff, Martin and Roberts, Keith and Walter, Peter},
   year={2013},
-  publisher={Garland Science}
+  publisher={Garland Science},
+  doi={}
+  
 }
 
 @book{krebs2013lewin,
@@ -265,5 +267,6 @@
   number={16},
   pages={2674--2695},
   year={2017},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.1111/febs.14150}
 }
