+++ paper/paper.bib
@@ -3,7 +3,7 @@
   author={Alberts, Bruce and Bray, Dennis and Hopkin, Karen and Johnson, Alexander D and Lewis, Julian and Raff, Martin and Roberts, Keith and Walter, Peter},
   year={2013},
   publisher={Garland Science},
-  doi={}
+  doi={10.1201/9781315815015}
   
 }
 
@@ -11,7 +11,8 @@
   title={Lewin's essential genes},
   author={Krebs, Jocelyn E and Lewin, Benjamin and Goldstein, Elliott S and Kilpatrick, Stephen T},
   year={2013},
-  publisher={Jones \& Bartlett Publishers}
+  publisher={Jones \& Bartlett Publishers},
+  doi={}
 }
 
 @article{o2013principles,
@@ -22,18 +23,20 @@
   number={7},
   pages={a010108},
   year={2013},
-  publisher={Cold Spring Harbor Lab}
+  publisher={Cold Spring Harbor Lab},
+  doi={10.1101/cshperspect.a010108}
 }
 
 @article{egan2005microreview,
-  title={MicroReview: Divided genomes: negotiating the cell cycle in prokaryotes with multiple chromosomes},
+  title={Divided genomes: negotiating the cell cycle in prokaryotes with multiple chromosomes},
   author={Egan, Elizabeth S and Fogel, Michael A and Waldor, Matthew K},
   journal={Molecular microbiology},
   volume={56},
   number={5},
   pages={1129--1138},
   year={2005},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.1111/j.1365-2958.2005.04622.x}
 }
 
 @article{frank2000oriloc,
@@ -44,7 +47,8 @@
   number={6},
   pages={560--561},
   year={2000},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
+  doi={10.1093/bioinformatics/16.6.560}
 }
 
 @article{gao2008ori,
@@ -55,14 +59,16 @@
   number={1},
   pages={79},
   year={2008},
-  publisher={BioMed Central}
+  publisher={BioMed Central},
+  doi={10.1186/1471-2105-9-79}
 }
 
 @book{beauchamp1979digital,
   title={Digital methods for signal analysis},
   author={Beauchamp, Kenneth G and Yuen, CK},
   year={1979},
-  publisher={Routledge}
+  publisher={Routledge},
+  doi={}
 }
 
 @article{shah2012nucleotide,
@@ -73,7 +79,8 @@
   number={1},
   pages={52--55},
   year={2012},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.biosystems.2011.09.003}
 }
 
 @article{shah2012origin,
@@ -84,7 +91,8 @@
   number={3},
   pages={142--144},
   year={2012},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.biosystems.2011.11.006}
 }
 
 @article{mrazek1998strand,
@@ -95,7 +103,8 @@
   number={7},
   pages={3720--3725},
   year={1998},
-  publisher={National Acad Sciences}
+  publisher={National Acad Sciences},
+  doi={10.1073/pnas.95.7.3720}
 }
 
 @article{lobry1996origin,
@@ -106,7 +115,8 @@
   number={5262},
   pages={745--746},
   year={1996},
-  publisher={American Association for the Advancement of Science}
+  publisher={American Association for the Advancement of Science},
+  doi={10.1126/science.272.5262.745}
 }
 
 @article{lobry2002asymmetric,
@@ -117,7 +127,8 @@
   number={10},
[…]
