+++ paper/paper.bib
@@ -4,16 +4,8 @@
   year={2013},
   publisher={Garland Science},
   doi={10.1201/9781315815015}
-  
 }
 
-@book{krebs2013lewin,
-  title={Lewin's essential genes},
-  author={Krebs, Jocelyn E and Lewin, Benjamin and Goldstein, Elliott S and Kilpatrick, Stephen T},
-  year={2013},
-  publisher={Jones \& Bartlett Publishers},
-  doi={}
-}
 
 @article{o2013principles,
   title={Principles and concepts of DNA replication in bacteria, archaea, and eukarya},
@@ -93,181 +85,6 @@
   year={2012},
   publisher={Elsevier},
   doi={10.1016/j.biosystems.2011.11.006}
-}
-
-@article{mrazek1998strand,
-  title={Strand compositional asymmetry in bacterial and large viral genomes},
-  author={Mr{\'a}zek, JAN and Karlin, Samuel},
-  journal={Proceedings of the National Academy of Sciences},
-  volume={95},
-  number={7},
-  pages={3720--3725},
-  year={1998},
-  publisher={National Acad Sciences},
-  doi={10.1073/pnas.95.7.3720}
-}
-
-@article{lobry1996origin,
-  title={Origin of replication of Mycoplasma genitalium},
-  author={Lobry, JR},
-  journal={Science},
-  volume={272},
-  number={5262},
-  pages={745--746},
-  year={1996},
-  publisher={American Association for the Advancement of Science},
-  doi={10.1126/science.272.5262.745}
-}
-
-@article{lobry2002asymmetric,
-  title={Asymmetric directional mutation pressures in bacteria},
-  author={Lobry, Jean R and Sueoka, Noboru},
-  journal={Genome biology},
-  volume={3},
-  number={10},
-  pages={research0058--1},
-  year={2002},
-  publisher={BioMed Central},
-  doi={}
-}
-
-@article{grigoriev1998analyzing,
-  title={Analyzing genomes with cumulative skew diagrams},
-  author={Grigoriev, Andrei},
-  journal={Nucleic acids research},
-  volume={26},
-  number={10},
-  pages={2286--2290},
-  year={1998},
-  publisher={Oxford University Press}
-}
-
-@article{zhang2003z,
-  title={The Z curve database: a graphic representation of genome sequences},
-  author={Zhang, Chun-Ting and Zhang, Ren and Ou, Hong-Yu},
-  journal={Bioinformatics},
-  volume={19},
-  number={5},
-  pages={593--599},
-  year={2003},
-  publisher={Oxford University Press}
-}
-
-@article{zhang2004identification,
-  title={Identification of replication origins in the genome of the methanogenic archaeon, Methanocaldococcus jannaschii},
-  author={Zhang, Ren and Zhang, Chun-Ting},
-  journal={Extremophiles},
-  volume={8},
-  number={3},
-  pages={253--258},
-  year={2004},
-  publisher={Springer}
-}
-
-@article{zhang2005identification,
-  title={Identification of replication origins in archaeal genomes based on the Z-curve method},
-  author={Zhang, Ren and Zhang, Chun-Ting},
-  journal={Archaea},
-  volume={1},
-  number={5},
-  pages={335--346},
-  year={2005},
-  publisher={Hindawi Publishing Corporation}
-}
-
-@article{luo2014ori,
-  title={Ori-Finder 2, an integrated tool to predict replication origins in the archaeal genomes},
-  author={Luo, Hao and Zhang, Chun-Ting and Gao, Feng},
-  journal={Frontiers in microbiology},
-  volume={5},
-  pages={482},
-  year={2014},
-  publisher={Frontiers}
-}
-
-@article{slonczewski2009microbiology,
[…]
