### 2019-07-21 #1585 @urmi-21
@csoneson Hi, I've added the DOIs to the references. Thanks.
