### 2019-07-23 urmi-21/ORIS#3 @holtgrewe
Installation instructions ambiguous
RE: https://github.com/openjournals/joss-reviews/issues/1589

A minor point.

There is no JAR to download, only a ZIP containing a JAR.
