### 2019-07-23 urmi-21/ORIS#4 @holtgrewe
Download button does not work
RE: https://github.com/openjournals/joss-reviews/issues/1589

On windows, instead the following message is displayed "Firefox is already running, but is not responding. The old Firefox process must be closed to open a new window."
