### 2019-07-23 urmi-21/ORIS#5 @holtgrewe
Target audience could be defined more explicitely
RE: https://github.com/openjournals/joss-reviews/issues/1589

Please specify more closely who the target audience is and what the prerequisites of knowledge are.
