### 2019-07-23 urmi-21/ORIS#9 @holtgrewe
Automated or manual tests
RE https://github.com/openjournals/joss-reviews/issues/1589

It is hard to verify that the functionality of the software works as described. First of all, not all functionality is described (cf. #8) but also there neither are automated/CI tests nor are there instructions for the functionalities to be tested.

I think that's a petty. The software appears to provide a large number of features but documentation and tests appear to be a bit lower than the standard of the features themselves.
