### 2019-07-23 urmi-21/ORIS#10 @holtgrewe
Please refine community guidelines
RE https://github.com/openjournals/joss-reviews/issues/1589

> Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

While point 2&3 are solved by the README statement "Please report all the bugs through GitHub here.", the first appears to be not addressed.
