### 2019-07-24 urmi-21/ORIS#4 @urmi-21
Thanks for reporting this issue. I have not been able to reproduce this error on my Windows (or MAC and Linux) environment. I think the problem may be due to firefox (please see this https://support.mozilla.org/en-US/kb/firefox-already-running-not-responding).
Please give this another try after killing the firefox process using the Windows Task Manager.
