+++ README.md
@@ -5,6 +5,7 @@
 ## Introduction
 ORIS is a novel tool, written in JAVA, that lets the user interactively explore the whole genome sequence data using a number of computational methods and charts. ORIS is well suited for finding the origin of replication sites across bacteria, archaea and to some extent eukaryotes. ORIS successfully predicted the origin of replication sites in the genome of *Plasmodium falciparum* (Agarwal, M., Bhowmick, K., Shah, K., Krishnamachari, A., & Dhar, S. K. (2017). Identification and characterization of ARS‐like sequences as putative origin (s) of replication in human malaria parasite Plasmodium falciparum. The FEBS journal, 284(16), 2674-2695.).
 
+ORIS is developed particularly for biologists and researchers who are working in the area of DNA replication. ORIS allows users with little or no programming background to interactively explore whole genome sequences and identify the putative origin of replication sites in the genome of interest. All the computational methods implemented in ORIS are accessible through a simple and intuitive GUI. We hope this will help biologists comprehensively analyze their DNA sequences.
 We have described applicability of our tool, ORIS, in a case study available as a supplementary document (https://github.com/urmi-21/ORIS). The method details could also be found in the supplementary document. A user guide is available from https://github.com/urmi-21/ORIS.
 
 ## Citation
