+++ CONTRIBUTING.md
@@ -1,0 +1,6 @@
+# Contributing
+
+To contribute to ORIS, please fork this repository, make changes and then send a pull request.
+Before contributing to this project please discuss the suggested changes via issues or email.
+
+
