### 2019-07-24 urmi-21/ORIS#10 @urmi-21
To address this, I've added CONTRIBUTING.md.
