### 2019-07-24 urmi-21/ORIS#9 @urmi-21
Hi, I agree that there are no automated test sets for ORIS.
But I have pointed out a manual way which lets the user test (or validate) the results produced by ORIS (see https://github.com/urmi-21/ORIS/tree/master/test). I have suggested to use ORIS’ automatic prediction module. This executes the skew, correlation, entropy and motif searching methods. The output produced then could be validated against known or expected output.

Further, in the case study performed, we have comprehensively tested ORIS using bacterial and archeal genome. The results are compared to previously experimentally validated and published results. We also compared ORIS with other software and tools and the results demonstrate that ORIS is able to find the correct origin of replication sequences in the given genomes.
Another, example for validation of ORIS’ result is the work published by Agarwal et. al. (https://www.ncbi.nlm.nih.gov/pubmed/28644560). In this work ORIS predicted putative origin of replication sites in the genome of Plasmodium falciparum. The computational results were later experimentally verified.

I hope this is sufficient evidence to confirm the functionality of the software and judge the correctness of the results produced by ORIS. I have also updated the test document (see https://github.com/urmi-21/ORIS/tree/master/test) to reflect this information.

Please let me know what you think.
