### 2019-07-24 urmi-21/ORIS#9 @holtgrewe
IMO it is sufficient but not obvious to find.

Please link to the tests directory and the contained README from the main README.
