### 2019-07-24 urmi-21/ORIS#10 @holtgrewe
Ah, sorry for closing this prematurely.

I think I don't understand the technical prerequisites for contributing. Do I have to use some IDE? How to build this? What are library dependencies here?

This should probably be part of the CONTRIBUTING file, documenting in text without figures is sufficient. It would be great if you could walk a potential contributor from cloning the repository to building and running the actual app.
