+++ README.md
@@ -8,24 +8,6 @@ ORIS is a novel tool, written in JAVA, that lets the user interactively explore 
 ORIS is developed particularly for biologists and researchers who are working in the area of DNA replication. ORIS allows users with little or no programming background to interactively explore whole genome sequences and identify the putative origin of replication sites in the genome of interest. All the computational methods implemented in ORIS are accessible through a simple and intuitive GUI. We hope this will help biologists comprehensively analyze their DNA sequences.
 We have described applicability of our tool, ORIS, in a case study available as a supplementary document (https://github.com/urmi-21/ORIS). The method details could also be found in the supplementary document. A user guide is available from https://github.com/urmi-21/ORIS.
 
-## Citation
-Coming Soon...
-
-
-## Examples
-
-
-### Reading Genome Data
-![alt text](https://github.com/urmi-21/ORIS/blob/master/figures/open.gif)
-
-
-### Plotting GC-Skew
-![alt text](https://github.com/urmi-21/ORIS/blob/master/figures/GCskew.gif)
-
-### Searching Motifs
-![alt text](https://github.com/urmi-21/ORIS/blob/master/figures/DnaASearch.gif)
-
-
 ## Getting Started
 
 ### Prerequisites
@@ -44,6 +26,26 @@ User guide describing the functionality is available at: https://github.com/urmi
 
 ### Methods and Case Study
 A document describing ORIS' methods along with a case study is available here: https://github.com/urmi-21/ORIS/blob/master/ORIS_SI.pdf
+
+## Citation
+Coming Soon...
+
+
+## Examples
+
+
+### Reading Genome Data
+![alt text](https://github.com/urmi-21/ORIS/blob/master/figures/open.gif)
+
+
+### Plotting GC-Skew
+![alt text](https://github.com/urmi-21/ORIS/blob/master/figures/GCskew.gif)
+
+### Searching Motifs
+![alt text](https://github.com/urmi-21/ORIS/blob/master/figures/DnaASearch.gif)
+
+
+
 
 ## Developers
 
