### 2019-07-25 a1c5a5e
Update CONTRIBUTING.md
