+++ CONTRIBUTING.md
@@ -4,3 +4,7 @@ To contribute to ORIS, please fork this repository, make changes and then send a
 Before contributing to this project please discuss the suggested changes via issues or email.
 
 
+# Getting Started
+## Dependencies
+jmathplot
+
