### 2019-07-25 32cbdce
Update CONTRIBUTING.md
