### 2019-07-25 7224c14
Update CONTRIBUTING.md
