### 2019-07-25 c45182a
Update CONTRIBUTING.md
