### 2019-07-25 urmi-21/ORIS#10 @urmi-21
Thanks for the suggestions. I have updated the CONTRIBUTING.md file.
