### 2019-07-25 #1589 @holtgrewe
@csoneson Is checking that the software is deposited in figshare/zeonodo etc. part of the editor tasks or might this be missing from the checklist?
