### 2019-07-25 urmi-21/ORIS#4 @holtgrewe
Hm, I'm seeing now that if firefox is closed, firefox is opened in an older version (probably the system wide installation vs. the automatically updated one in the user directory). I'm closing this now but maybe there is a more robust way of opening a URL than the one you are using?
