### 2019-07-25 urmi-21/ORIS#9 @holtgrewe
OK, I have tested the program now and it generates output similar to the GIF in the README. I trust that this makes sense in the application domain (that I have no knowledge about) ;)

Reading through the user guide, it might make sense to include some more information on how to interpret the results. If "wet-lab biologists" are to use the software they might welcome help of how to interpret the "dry-lab" results that your program gives.
