### 2019-07-25 #1589 @holtgrewe
OK, the software is runnable and does what it's supposed to do by the user guide and the GIFs on the home page. I cannot speak for content but trust the author that it's the same functionality they used in other published and reviewed research. @urmi-21 I have a number of additional (software engineering) suggestions that I'll open as a separate ticket in the repository.
