### 2019-07-25 #1589 @csoneson
@holtgrewe The final zenodo archive is generated at the time when the submission is accepted. So there doesn't have to be one at this time.
