### 2019-07-26 #1589 @jdeligt
The submission is looking good and thanks to the suggestions by @csoneson and @holtgrewe and the subsequent changes by @urmi-21 most checks are looking excellent. I'll be testing it on some real data over the weekend on a cluster and locally to ensure it can install and run on a diverse architecture. If I encounter and deal breakers I'll link the issues here. Any suggestions or minor points I'll report as issues in the repo but won't let them hold up the progress.
