### 2019-07-28 urmi-21/ORIS#13 @jdeligt
Please include build instuctions
Please add build instructions for those users that would like to check out the code and build themselves. A precompiled jar is subject to injections, i.e. not guaranteed to be the source of the repo.
