### 2019-07-28 urmi-21/ORIS#14 @jdeligt
Please log actions of user
In the light of reproducible research it would be good to log the actions performed through the GUI. Having to keep track of all settings/action will require a lot of documentation/logging by the user to be able to use the obtained results in their research.
