### 2019-07-28 #1589 @jdeligt
The software can be download and installed according to the instructions. I would personally prefer build instructions to be specified for those users that would like to check out the code and build themselves. A precompiled jar is subject to injections, i.e. not guaranteed to be the source of the repo, I've raised an issue #13 for this. I understand the target audience is less code savvy but I think this is good practice for OSS.

For my second point (#14) I would like the input from @csoneson . In the light of reproducible research I looked for options to log the actions performed through the GUI. I was unable to find it and the default behavior doesn't log it to STDout or a file. @urmi-21 if I missed it sorry please point it out to me. If this is indeed the case I feel it will require a lot of documentation / logging by the user to be able to use the obtained results in their research. Thereby limiting the usefulness and ease of use of the software. This is not part of the default checklist and as such would like a second opinion if we want to hold up the publication based on this or not.
