### 2019-07-28 urmi-21/ORIS#13 @urmi-21
Hi, the building instructions here: https://github.com/urmi-21/ORIS/blob/master/CONTRIBUTING.md
