### 2019-07-28 urmi-21/ORIS#13 @jdeligt
Hi, thanks for pointing that out. Not where I expected them. Could you refer to it from the install instructions or do you want it there so only contributors have to see it?
