### 2019-07-28 urmi-21/ORIS#13 @urmi-21
I have added a building section in the readme which points to the contributing file
