### 2019-07-28 urmi-21/ORIS#14 @urmi-21
Hi,
ORIS doesn't log the user actions performed via GUI. The user is expected to log information on their own when they find useful results (e.g., exporting the location of origin using auto-correlation method) and the results could be easily reproduced if the user provides the data, the methods and the parameters used.

ORIS is developed for interactive exploration of whole-genome sequences and it provides specialized methods for finding the origin of replication sequences. This process of exploratory data analysis is usually non-linear and may take several steps until a conclusion is reached. This can generate very big logs and most of the details in the log may be of little importance. However, the results which matter to the user could be easily exported when required. We leave it to the user to best describe/document the results and these results may be easily reproduced again using the same data and the methods.

Logging all the user actions may enable some advance users to keep track of the actions performed but I think ORIS’ ability to export results along with user documentation is sufficient for reproducibility.

If one wants to rely on logs then again users need to take an extra step to mine the useful parts in the log file. I agree that this could be complementary to what ORIS can do right now but not having this doesn’t limit the applicability of the software.

I would also like to say that the GUI itself is designed so that the software is easy-to-use for the people who are not familiar with coding.

The settings, however, are saved in a file called _settings.cfg_ which is located in the same directory as the .jar file.
