### 2019-07-29 #1589 @jdeligt
@csoneson I've closed  urmi-21/ORIS#13 and check the last box since the software meets the basic requirements of JOSS. I'm leaving urmi-21/ORIS#14 since I personally feel it does somewhat limit the usefullness in science especially when dealing with non coders they might not realise how important a specifc setting or algorithm might influence their findings. But that doesn't mean it is not a usefull piece of software and as such I am okay with progressing the publication.
