### 2019-07-29 urmi-21/ORIS#14 @jdeligt
Thank you for the response and explaining the rationale. And just to be clear I think it is great that you went throught the effort of making a GUI to enable people to do these types of analysis in a more intuative and interactive way.

However many interactive/online tools do log user actions and some can even reconstruct the path to a figure even when there is al lot og back and forth in settings/algorithms involved. Yes this is quite some work but it does allow the user to download a log of commands/settings that resulted in their fiigures/conclusions. This is great for publishing and in some cases allows re-using of certain steps without all the clicks.

It might be a matter of perspective/opnion but I would argue that these logs are especially relevant to individuals not familiar with coding. They might not be as aware of the impact the choice of profile search/algorithm might have on their conclusions. It's our resposibility as developers to make them aware of these and provide them with ways of seeing what is happening.

Please understand that this is not a ctitique on the usefullness of your GUI but to a common trend in interactive anlysis. These types of analysis are powerfull because the put tools in the hands of the people with the biological questions but they suffer from reproducibilty issues.

I'll leave the issue open so that users can be aware of the fact that they will have to do their own regious logging and decide for themselves wether this is an issue for them or not.
