### 2019-07-29 #1589 @csoneson
@urmi-21 - I agree with @jdeligt about the importance of somehow tracking the user's actions in the interest of reproducibility as well as record keeping. One option would be to store a snapshot of the settings when a file is saved (perhaps into a separate file, with a suffix compared to the main output file). Some of the text files output by ORIS already contain some such information, like the window size. Storing the value of all the settings (including the name of the input file) should allow the user to reproduce (and remember) the analyses going into a saved text file or plot.
