### 2019-07-29 #1589 @csoneson
@urmi-21 a couple of additional small points:
- [ ] The software expects a single-sequence fasta file - I think it would be helpful if an error was raised if the user provides a multi-fasta file (i.e., if there are multiple > characters in the file). Otherwise the IDs of the following sequences will be treated as part of the sequence.
- [ ] Why is X used instead of the standard IUPAC code N to represent "any" base?
- [ ] There are several consistent (and some inconsistent) typos in the software and documentation. E.g., "increament", "sequnce", "gaped" - please have another round through and correct these
