### 2019-07-29 eeee938
show error for multi fasta files
