### 2019-07-29 c0e5798
Update UserGuide.java
