+++ bioapp/src/bioapp/UserGuide.java
@@ -8,11 +8,12 @@ import javax.swing.ImageIcon;
 
 /**
  *
- * @author jnu
- * This class is an inbuilt user guide
+ * @author jnu This class is an inbuilt user guide
  */
 public class UserGuide extends javax.swing.JFrame {
-ImageIcon img = new ImageIcon(getClass().getClassLoader().getResource("images/icons/orislogo.png"));
+
+    ImageIcon img = new ImageIcon(getClass().getClassLoader().getResource("images/icons/orislogo.png"));
+
     /**
      * Creates new form UserGuide
      */
@@ -23,7 +24,7 @@ ImageIcon img = new ImageIcon(getClass().getClassLoader().getResource("images/ic
         this.setIconImage(img.getImage());
         this.setDefaultCloseOperation(DISPOSE_ON_CLOSE);
         this.setResizable(false);
-        
+
     }
 
     /**
@@ -203,7 +204,7 @@ ImageIcon img = new ImageIcon(getClass().getClassLoader().getResource("images/ic
 
     private void jLabelskewMouseEntered(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_jLabelskewMouseEntered
         // TODO add your handling code here:
-        
+
     }//GEN-LAST:event_jLabelskewMouseEntered
 
     private void jLabelcorrMouseClicked(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_jLabelcorrMouseClicked
@@ -213,54 +214,58 @@ ImageIcon img = new ImageIcon(getClass().getClassLoader().getResource("images/ic
 
     private void jLabelsearchMouseClicked(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_jLabelsearchMouseClicked
         // TODO add your handling code here:
-        jTextAreahelp.setText("Pattern search is helpful for finding motifs or regions of intereset in a given sequence.\nTo find a pattern, after reading file:\n1. Click pattern search\n2. Choose a type of search\n3. Choose appropriate parameters for search.\n\n\n Parameters for pattern search:\na. Reverse Complement: Check this option to also search the reverse complement sequnce.\nb. Sequence logo: Check this option to generate a sequence logo based on search results.\nc. Search Statics: Check this option if you want to view graphically, where the matches have been \nfound in the sequnce.\nd. Save results: Check this option to save search results into a .txt file\ne. Tolerance: select a tolerance value. This will allow mismatches, e.g. search with tolerance \nvalue=1 will allow 1 mismatch.");
+        jTextAreahelp.setText("Pattern search is helpful for finding motifs or regions of intereset in a given sequence.\nTo find a pattern, after reading file:\n1. Click pattern search\n2. Choose a type of search\n3. Choose appropriate parameters for search.\n\n\n Parameters for pattern search:\na. Reverse Complement: Check this option to also search the reverse complement sequence.\nb. Sequence logo: Check this option to generate a sequence logo based on search results.\nc. Search Statics: Check this option if you want to view graphically, where the matches have been \nfound in the sequence.\nd. Save results: Check this option to save search results into a .txt file\ne. Tolerance: select a tolerance value. This will allow mismatches, e.g. search with tolerance \nvalue=1 will allow 1 mismatch.");
     }//GEN-LAST:event_jLabelsearchMouseClicked
 
     private void jLabelwtmatrixMouseClicked(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_jLabelwtmatrixMouseClicked
         // TODO add your handling code here:
[…]
