+++ bioapp/src/bioapp/Form2.java
@@ -3,7 +3,7 @@
  * and open the template in the editor.
  */
 
-/*
+ /*
  * Form2.java
  *
  * Created on Mar 24, 2013, 11:55:01 PM
@@ -49,13 +49,12 @@ import javax.swing.text.*;
 
 /**
  *
- * @author urmi
- * This is the main display form for ORIS.
+ * @author urmi This is the main display form for ORIS.
  */
 public class Form2 extends javax.swing.JFrame {
 
-   public ImageIcon img = new ImageIcon(ImageIO.read(getClass().getResourceAsStream("/images/icons/orislogo.png")));
-   
+    public ImageIcon img = new ImageIcon(ImageIO.read(getClass().getResourceAsStream("/images/icons/orislogo.png")));
+
     int filereadflag;
     SimpleAttributeSet set;
     String filedata;
@@ -65,7 +64,7 @@ public class Form2 extends javax.swing.JFrame {
     static public char PRIORITY;
     static public String DEFAULT_OPEN_DIR;
     static public String DEFAULT_SAVE_DIR;
-    
+
     /*public BufferedImage getImageStream(String path){
         try {
             return ImageIO.read(getClass().getResourceAsStream(path));
@@ -79,7 +78,6 @@ public class Form2 extends javax.swing.JFrame {
         return getClass().getClassLoader().getResource(path);
        
     }*/
-
     /**
      * Creates new form Form2
      */
@@ -226,7 +224,7 @@ public class Form2 extends javax.swing.JFrame {
         jToolBar1.setFloatable(false);
         //jButtonopen.setIcon(new ImageIcon("images/icons/folder.png", "open"));
         jButtonopen.setIcon(new ImageIcon(getClass().getClassLoader().getResource("images/icons/folder.png"), "open"));
-                
+
         jButtondownload.setIcon(new ImageIcon(getClass().getClassLoader().getResource("images/icons/download.png"), "download"));
         jButtondownload.setIcon(new ImageIcon(getClass().getClassLoader().getResource("images/icons/download.png"), "download"));
         jButtonfind.setIcon(new ImageIcon(getClass().getClassLoader().getResource("images/icons/search.png"), "search"));
@@ -344,6 +342,7 @@ public class Form2 extends javax.swing.JFrame {
         jMenuItemsettings = new javax.swing.JMenuItem();
         jMenuItemabout = new javax.swing.JMenuItem();
         jMenuItemguide = new javax.swing.JMenuItem();
+        jMenuItem2 = new javax.swing.JMenuItem();
         jMenuItemjmath = new javax.swing.JMenuItem();
 
         jMenuItem1.setText("jMenuItem1");
@@ -921,6 +920,14 @@ public class Form2 extends javax.swing.JFrame {
         });
         jMenu2.add(jMenuItemguide);
 
+        jMenuItem2.setText("GitHub");
+        jMenuItem2.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                jMenuItem2ActionPerformed(evt);
+            }
+        });
+        jMenu2.add(jMenuItem2);
+
         jMenuItemjmath.setText("Jmathplot");
         jMenuItemjmath.addActionListener(new java.awt.event.ActionListener() {
             public void actionPerformed(java.awt.event.ActionEvent evt) {
@@ -1208,18 +1215,16 @@ public class Form2 extends javax.swing.JFrame {
                     offset = pos[i] / 70;
                     // System.out.printf("offset=%d\n", offset);
                     try {
-                        
-                        String thisOS= System.getProperty("os.name");
-                        if(thisOS.contains("Windows")){
+
+                        String thisOS = System.getProperty("os.name");
+                        if (thisOS.contains("Windows")) {
                             //for windows
[…]
