### 2019-07-29 #1589 @urmi-21
@jdeligt @csoneson Thanks for your comments and suggestions regarding tracking user actions. I can see that it will be helpful for users and that will make the software more user-friendly. But if not necessary for this publication I would like to keep this update for future releases. Thanks, @jdeligt for opening this issue in the ORIS repository.
