### 2019-07-29 #1589 @urmi-21
>
>
> @urmi-21 a couple of additional small points:
>
>     * [ ]  The software expects a single-sequence fasta file - I think it would be helpful if an error was raised if the user provides a multi-fasta file (i.e., if there are multiple > characters in the file). Otherwise the IDs of the following sequences will be treated as part of the sequence.
>
>     * [ ]  Why is X used instead of the standard IUPAC code N to represent "any" base?
>
>     * [ ]  There are several consistent (and some inconsistent) typos in the software and documentation. E.g., "increament", "sequnce", "gaped" - please have another round through and correct these

@csoneson Thanks for your comments.

1. I have addressed the first issue and now an error message is thrown if a multi fasta file is read.

2. X is used instead of N to actually distinguisg N from other bases (A, G, C, T). While searching user can choose to explicitly match N or match X, which will match N and other bases.
E.g.
Target seq    ATGCGANTG
Searching NTG will give one match
Searching XTG will give two matches (ATG and NTG)

3. Sorry for not closely checking the spellings before. I have fixed the typos now.
