### 2019-07-30 #1589 @csoneson
> @jdeligt @csoneson Thanks for your comments and suggestions regarding tracking user actions. I can see that it will be helpful for users and that will make the software more user-friendly. But if not necessary for this publication I would like to keep this update for future releases. Thanks, @jdeligt for opening this issue in the ORIS repository and thanks for reviewing my submission. I really appreciate your efforts

@urmi-21 - I am willing to move forward with this submission without the tracking of user actions, leaving it to future releases; however, in that case I would ask you to add a statement in the paper clarifying that this feature is currently not implemented, so that users are aware of what they have to keep track of manually in order to achieve reproducibility.
