### 2019-07-30 #1589 @holtgrewe
@csoneson I share @jdeligt's concerns regarding reproducible (as always with UI tools) and also reproducing builds. The first is appropriately tackled IMO once the author adds clarification to the manual. The second is addressed in an issue already.

Actually, I'd like to see a change in JOSS policy requiring automated builds (not CI but using a common, non-IDE build system approach, e.g., at least Ant or Maven for Java as in this case) and nudging towards reproducibility/tracing. However, that's not for me to decide.

Overall, yes, I think this follows the current JOSS guidelines appropriately and I thus have no concerns regarding it fulfilling JOSS requirements.
