### 2019-07-31 #1589 @jdeligt
@csoneson thanks for the input and suggesting the addtion to the manual.  As @holtgrewe points out build systems aren't currently coverd and there is room for improvement there. The tool and publication meet the guidelines and look forward to the authors continuing to work on the reproducibility. As such I'm happy to progress the publication.
