### 2019-07-31 144726b
Update paper.md
