### 2019-07-31 9715b3c
Update paper.md
