+++ paper/paper.md
@@ -48,7 +48,7 @@ ORIS is a novel tool that lets the user interactively explore the whole genome s
 
 We have tested and validated our tool, ORIS, by performing a case study, available as a supplementary document (https://github.com/urmi-21/ORIS). The method details are also included in the supplementary document. A user guide is available from https://github.com/urmi-21/ORIS.
 
-ORIS is developed particularly for biologists and researchers who are working in the area of DNA replication. ORIS allows users with little or no programming background to interactively explore whole genome sequences and identify the putative origin of replication sites in the genome of interest. An advantage of using ORIS is that all the computational methods implemented in ORIS are accessible through a simple and intuitive GUI. A potential drawback of using the GUI is that the user has to manually document/log the metadata associated with the results (e.g., the method used, method parameters, and input files). We plan to implement automatic logging in future releases of ORIS so that this information is easily recorded along with the results, which will make it easier to reproduce the results while minimizing user efforts.
+ORIS is developed particularly for biologists and researchers who are working in the area of DNA replication. ORIS allows users with little or no programming background to interactively explore whole genome sequences and identify the putative origin of replication sites in the genome of interest. An obvious advantage of using ORIS is that all the computational methods implemented in ORIS are accessible through a simple and intuitive GUI but the user has to manually document/log the metadata associated with the results (e.g., the method used, method parameters, and input files). We plan to implement automatic logging in future releases of ORIS so that this information is easily recorded along with the results, which will make it, even more, easier to reproduce the results while minimizing user efforts.
 
 
 # Acknowledgements
