### 2019-07-31 cecc589
Minor fixes
