+++ paper/paper.bib
@@ -8,9 +8,9 @@
 
 
 @article{o2013principles,
-  title={Principles and concepts of DNA replication in bacteria, archaea, and eukarya},
+  title={Principles and concepts of {DNA} replication in bacteria, archaea, and eukarya},
   author={O’Donnell, Michael and Langston, Lance and Stillman, Bruce},
-  journal={Cold Spring Harbor perspectives in biology},
+  journal={Cold Spring Harbor Perspectives in Biology},
   volume={5},
   number={7},
   pages={a010108},
@@ -22,7 +22,7 @@
 @article{egan2005microreview,
   title={Divided genomes: negotiating the cell cycle in prokaryotes with multiple chromosomes},
   author={Egan, Elizabeth S and Fogel, Michael A and Waldor, Matthew K},
-  journal={Molecular microbiology},
+  journal={Molecular Microbiology},
   volume={56},
   number={5},
   pages={1129--1138},
@@ -44,9 +44,9 @@
 }
 
 @article{gao2008ori,
-  title={Ori-Finder: a web-based system for finding oriC s in unannotated bacterial genomes},
+  title={Ori-Finder: a web-based system for finding oriCs in unannotated bacterial genomes},
   author={Gao, Feng and Zhang, Chun-Ting},
-  journal={BMC bioinformatics},
+  journal={BMC Bioinformatics},
   volume={9},
   number={1},
   pages={79},
@@ -66,7 +66,7 @@
 @article{shah2012nucleotide,
   title={Nucleotide correlation based measure for identifying origin of replication in genomic sequences},
   author={Shah, Kushal and Krishnamachari, Annangarachari},
-  journal={BioSystems},
+  journal={Biosystems},
   volume={107},
   number={1},
   pages={52--55},
@@ -88,9 +88,9 @@
 }
 
 @article{agarwal2017identification,
-  title={Identification and characterization of ARS-like sequences as putative origin (s) of replication in human malaria parasite Plasmodium falciparum},
+  title={Identification and characterization of {ARS}-like sequences as putative origin(s) of replication in human malaria parasite {P}lasmodium falciparum},
   author={Agarwal, Meetu and Bhowmick, Krishanu and Shah, Kushal and Krishnamachari, Annangarachari and Dhar, Suman Kumar},
-  journal={The FEBS journal},
+  journal={The FEBS Journal},
   volume={284},
   number={16},
   pages={2674--2695},
