+++ paper/paper.md
@@ -31,9 +31,9 @@ bibliography: paper.bib
 # Summary
 
 A remarkable yet very precise event that takes place within the cell, is called DNA replication or chromosome copy [@alberts2013essential].
-The way it is done in eukaryotes is vastly different from prokaryotes yet something is common as well. The origin of replication is a small segment in the genome at which the DNA replication process is initiated. Bacterial genomes generally have a single origin of DNA replication whereas archea and eukaryotes have multiple replication sites [@o2013principles; @egan2005microreview].
+The way it is done in eukaryotes is vastly different from prokaryotes yet something is common as well. The origin of replication is a small segment in the genome at which the DNA replication process is initiated. Bacterial genomes generally have a single origin of DNA replication whereas archaea and eukaryotes have multiple replication sites [@o2013principles; @egan2005microreview].
 
-Experimental studies carried out on the model organism *S. cerevisiae* brought out greater insights in understanding the complex replication process. A typical origin sequence in *S. cerevisiae* has a DNA segment called Autonomous Replication Sequence (ARS) and this further has an essential A element and multiple B elements and all these are required for proper origin function or firing. This kind of organizational structure may be present in other genomes but is not very obvious. Uncovering such unknown structures in the genome can improve our understanding of DNA replication and the origin of replication sites.
+Experimental studies carried out on the model organism *S. cerevisiae* brought greater insights in understanding the complex replication process. A typical origin sequence in *S. cerevisiae* has a DNA segment called Autonomous Replication Sequence (ARS) and this further has an essential A element and multiple B elements and all these are required for proper origin function or firing. This kind of organizational structure may be present in other genomes but is not very obvious. Uncovering such unknown structures in the genome can improve our understanding of DNA replication and the origin of replication sites.
 
 Hence, software tools should be developed incorporating many such useful contextual information for sequence analysis, with the hope that biologists will benefit and it may provide them a helping hand to search for origin like sequences in any genome of their choice. 
 
