### 2019-07-31 #1589 @csoneson
@urmi-21 - I have made a PR with a couple of formatting fixes to the paper. Please review and merge, and generate a new pdf proof.

In addition to that, the next step is to make a tagged release of your software, and create a Zenodo archive. Please make sure that the Zenodo archive has exactly the same title and authors as the paper. Once that's done, please report back here with the DOI of the Zenodo archive.
