### 2019-08-01 3db53c1
Update paper.md
