+++ paper/paper.md
@@ -20,11 +20,11 @@ authors:
 affiliations:
  - name: School of Computational and Integrative Sciences, Jawaharlal Nehru University, New Delhi, India
    index: 1
- - name: Department of Electrical Engineering and Bharti School of Telecommunication Technology and Management, IIT Delhi, India
+ - name: Department of Electrical Engineering and Computer Science, Indian Institute of Science Education and Research (IISER), Bhopal - 462066, Madhya Pradesh, India
    index: 2
  - name: Special centre for Molecular Medicine, Jawaharlal Nehru University, New Delhi, India
    index: 3
-date: 18 July 2019
+date: 31 July 2019
 bibliography: paper.bib
 ---
 
