### 2019-08-01 #1589 @csoneson
@openjournals/joss-eics - I believe we are ready to have this paper accepted!
