### 2019-08-01 #1589 @labarba
The *title* of the release doesn't match the tag (says "v1.0" while tag is "v1.0.2")
https://github.com/urmi-21/ORIS/releases
