### 2019-08-01 #1589 @labarba
@urmi-21 — I wonder if you would be amenable to making some final improvements in your paper. It felt like a bit of a slog to have to read half a page of background before getting to a statement about what the software does. Could you look into placing a few sentences about the software right at the top of your summary? I feel like the first portion of the _last_ paragraph, with slight edits, would make a good opening. The reader wants to know at the start: *Who is this software for? What does it do?*

Also, could you remove the link to the repository that appears twice on the second page? The link to the repo will be part of the first-page margin decorators, so you can simply mention it in the text body without adding the URL.

And while you're editing, a final suggestion I have is to try to remove some unnecessary passive constructions that make sentences long and less readable.
