+++ paper/paper.bib
@@ -98,3 +98,14 @@
   publisher={Wiley Online Library},
   doi={10.1111/febs.14150}
 }
+
+@article{grimwade2017targeting,
+  title={Targeting the bacterial orisome in the search for new antibiotics},
+  author={Grimwade, Julia E and Leonard, Alan C},
+  journal={Frontiers in microbiology},
+  volume={8},
+  pages={2352},
+  year={2017},
+  publisher={Frontiers},
+  doi={10.3389/fmicb.2017.02352}
+}
