### 2019-08-02 c6a7d28
Update paper.md
