+++ paper/paper.md
@@ -30,25 +30,16 @@ bibliography: paper.bib
 
 # Summary
 
-A remarkable yet very precise event that takes place within the cell, is called DNA replication or chromosome copy [@alberts2013essential].
-The way it is done in eukaryotes is vastly different from prokaryotes yet something is common as well. The origin of replication is a small segment in the genome at which the DNA replication process is initiated. Bacterial genomes generally have a single origin of DNA replication whereas archaea and eukaryotes have multiple replication sites [@o2013principles; @egan2005microreview].
+The origin of replication is a small segment in the genome at which the DNA replication process is initiated [@alberts2013essential]. Bacterial genomes generally have a single origin of DNA replication whereas archaea and eukaryotes have multiple replication sites [@o2013principles; @egan2005microreview]. The accurate identification of the origin of replication sequences is important to understand the complex replication process and could also provide new effective drug targets [@grimwade2017targeting].
 
-Experimental studies carried out on the model organism *S. cerevisiae* brought greater insights in understanding the complex replication process. A typical origin sequence in *S. cerevisiae* has a DNA segment called Autonomous Replication Sequence (ARS) and this further has an essential A element and multiple B elements and all these are required for proper origin function or firing. This kind of organizational structure may be present in other genomes but is not very obvious. Uncovering such unknown structures in the genome can improve our understanding of DNA replication and the origin of replication sites.
+Here, we present a standalone software tool, ORIS, for finding the origin of replication sites in the genome. ORIS is a novel tool that lets the user interactively explore the whole genome sequence data using several computational methods and charts. ORIS is developed particularly for biologists and researchers who are working in the area of DNA replication. ORIS allows users with little or no programming background to interactively explore whole genome sequences and identify the putative origin of replication sites in the genome of interest.
 
-Hence, software tools should be developed incorporating many such useful contextual information for sequence analysis, with the hope that biologists will benefit and it may provide them a helping hand to search for origin like sequences in any genome of their choice. 
+ORIS is well suited for origin finding across bacteria, archaea and to some extent eukaryotes. The origin of replication sites in *Plasmodium falciparum* were computationally predicted using ORIS and experimentally verified [@agarwal2017identification].
 
-There are a few online tools/databases which may help in finding the origin of replication sites for a few species [@frank2000oriloc; @gao2008ori]. These tools are limited by the type of computational approaches they take and are not interactive.
+There are a few online tools/databases which may help in finding the origin of replication sites for a few species [@frank2000oriloc; @gao2008ori]. These tools are limited by the type of computational approaches they take and are not interactive. Whereas, ORIS is a comprehensive software suite that can take any genome as input. It can perform context-based analysis, which includes DSP (digital signal processing) based correlation measures, all types of skew measures and entropy-based methods for predicting the origin of replication [@beauchamp1979digital; @shah2012nucleotide]. Features such as pattern matching, generating sequence logo, DNA bending profiles, and composition statistics are also available as part of the analysis. 
 
-Here, we present a standalone software tool, ORIS, for interactive analysis of prokaryotic genomes to find the origin of replication sites in the genome.
-An effort is made to develop a software suite that does context-based analysis which includes DSP (digital signal processing) based correlation measures, all types of skew measures and entropy-based methods for predicting the origin of replication [@beauchamp1979digital; @shah2012nucleotide].
-Also, features such as pattern matching, generating sequence logo, finding entropy distribution along the length of the genome, DNA bending profiles, and composition statistics are available as part of the analysis. 
-
[…]
