+++ paper/paper.md
@@ -39,6 +39,7 @@ ORIS is well suited for origin finding across bacteria, archaea and to some exte
 There are a few online tools/databases which may help in finding the origin of replication sites for a few species [@frank2000oriloc; @gao2008ori]. These tools are limited by the type of computational approaches they take and are not interactive. Whereas, ORIS is a comprehensive software suite that can take any genome as input. It can perform context-based analysis, which includes DSP (digital signal processing) based correlation measures, all types of skew measures and entropy-based methods for predicting the origin of replication [@beauchamp1979digital; @shah2012nucleotide]. Features such as pattern matching, generating sequence logo, DNA bending profiles, and composition statistics are also available as part of the analysis. 
 
 An obvious advantage of using ORIS is that all the computational methods implemented in ORIS are accessible through a simple and intuitive GUI but the user has to manually document/log the metadata associated with the results (e.g., the method used, method parameters, and input files). We plan to implement automatic logging in future releases of ORIS so that this information is easily recorded along with the results, which will make it, even more, easier to reproduce the results while minimizing user efforts.
+
 We have tested and validated our tool by performing a case study (available as a supplementary document from ORIS’ GitHub repository). The method details are also included in the supplementary document. A user guide is available from ORIS’ GitHub repository.
 
 
