### 2019-08-02 #1589 @urmi-21
@csoneson Thank you for accepting the paper.
@labarba Thank you for the comments. I have fixed the tag issue. the current release in now v1.0 this is also reflected in the zenodo archive. The doi is the same i.e.

10.5281/zenodo.3357162

I have reformatted the paper as suggested. I hope it's acceptable now.
