### 2019-08-02 #1589 @labarba
I'm confused: the release number goes _back_ from v1.0.2 to v1.0?
