### 2019-08-02 #1589 @urmi-21
>
>
> I'm confused: the release number goes _back_ from v1.0.2 to v1.0?

Sorry. Yes, it might be confusing. I have made it so that it's consistent with what is mentioned in the software. I would like this release to be v1.0. I have deleted the other releases/tags from GitHub.
