### 2019-08-02 #1589 @labarba
Huge thanks to our editor: @csoneson, and the reviewers: @jdeligt, @holtgrewe — your contribution to JOSS is greatly appreciated 🙏
